major refactoring and re-organizing
