+++ utils.py
@@ -1,0 +1,8 @@
+def fetch_readme():
+    try:
+        with open('README.md', 'r') as file:
+            return file.read()
+    except FileNotFoundError:
+        return "README.md file not found."
+    except Exception as e:
+        return f"An error occurred: {e}"
